# Return ReadOnlySequence slice as value instead of single-segment span