--- a/src/System.Text.JsonLab/System/Text/Json/InternalNumberResult.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/InternalNumberResult.cs
@@ -1,0 +1,13 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Text.JsonLab
+{
+    internal enum InternalNumberResult
+    {
+        Success,
+        OperationIncomplete,
+        NeedMoreData
+    }
+}