--- a/src/System.Text.JsonLab/System/Text/Json/Utf8JsonReader.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/Utf8JsonReader.cs
@@ -348,10 +348,6 @@ namespace System.Text.JsonLab
             return _isSingleSegment ? ReadSingleSegment() : ReadMultiSegment();
         }
 
-        public void Dispose()
-        {
-        }
-
         public void Skip()
         {
             if (TokenType == JsonTokenType.PropertyName)