**Changes:**
- Removed use of ArrayPool and the Dispose method
- Added `ValueSequence` and `IsValueMultiSegment` properties
- Split implementation into single-segment and multi-segment code paths again

**Results:**
- Multi-segment case regressed by 4-8%
- Single-segment case regressed by 8-14%
- Multi-segment input (sequence) is 5-20% slower than single-segment input (span)

**TODO:**
- ~Not all tests are expected to pass yet (particularly the error/invalid cases and partial data).~
- ~Fix SkipComment and ConsumeComment code paths.~
- ~Fix tests to use the right valuespan/valuesegment.~
- ~Revert benchmark test changes.~
- Optimize multi-segment string parsing (use vectorized IndexOf instead of a regular for-loop).

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17763
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-009706
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT


```

**Before:**

|                                            Method | IsDataCompact |   TestCase |          Mean |         Error |        StdDev |        Median | Scaled | ScaledSD | Allocated |
|-------------------------------------------------- |-------------- |----------- |--------------:|--------------:|--------------:|--------------:|-------:|---------:|----------:|
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicJson** |     **505.07 ns** |      **8.833 ns** |      **8.262 ns** |     **502.91 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  BasicJson |     746.14 ns |     12.845 ns |     10.726 ns |     741.89 ns |   1.48 |     0.03 |       0 B |
|                                                   |               |            |               |               |               |               |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **HelloWorld** |      **94.05 ns** |      **1.898 ns** |      **3.657 ns** |      **92.31 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True | HelloWorld |     303.80 ns |      4.607 ns |      4.310 ns |     304.31 ns |   3.23 |     0.13 |       0 B |
|                                                   |               |            |               |               |               |               |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json400B** |     **734.15 ns** |     **14.265 ns** |     **13.344 ns** |     **737.06 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |   Json400B |     984.16 ns |     10.303 ns |      9.134 ns |     988.65 ns |   1.34 |     0.03 |       0 B |
|                                                   |               |            |               |               |               |               |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **Json400KB** | **801,186.36 ns** | **14,926.742 ns** | **13,962.482 ns** | **797,099.65 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  Json400KB | 836,265.27 ns | 16,457.665 ns | 17,609.515 ns | 830,180.81 ns |   1.04 |     0.03 |       0 B |
|                                                   |               |            |               |               |               |               |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json40KB** |  **73,365.60 ns** |    **797.221 ns** |    **706.716 ns** |  **73,131.31 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |   Json40KB |  77,724.29 ns |  1,401.020 ns |  1,310.515 ns |  77,261.26 ns |   1.06 |     0.02 |       0 B |
|                                                   |               |            |               |               |               |               |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |    **Json4KB** |   **5,729.75 ns** |     **70.696 ns** |     **62.670 ns** |   **5,722.50 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |    Json4KB |   6,216.36 ns |     60.006 ns |     53.193 ns |   6,220.39 ns |   1.09 |     0.01 |       0 B |

|                Method |  TestCase | segmentSize |         Mean |      Error |     StdDev | Allocated |
|---------------------- |---------- |------------ |-------------:|-----------:|-----------:|----------:|
|  **MultiSegmentSequence** | **Json400KB** |        **4000** | **1,111.700 us** | **19.6317 us** | **17.4030 us** |       **0 B** |
| **SingleSegmentSequence** | **Json400KB** |           **?** | **1,080.178 us** |  **6.5580 us** |  **6.1344 us** |       **0 B** |
|  **MultiSegmentSequence** |  **Json40KB** |        **4000** |   **107.378 us** |  **1.3430 us** |  **1.1905 us** |       **0 B** |
| **SingleSegmentSequence** |  **Json40KB** |           **?** |   **104.726 us** |  **0.8713 us** |  **0.7724 us** |       **0 B** |
|  **MultiSegmentSequence** |   **Json4KB** |        **4000** |     **8.853 us** |  **0.1266 us** |  **0.1184 us** |       **0 B** |
| **SingleSegmentSequence** |   **Json4KB** |           **?** |     **8.832 us** |  **0.0734 us** |  **0.0651 us** |       **0 B** |

**After:**

|                                            Method | IsDataCompact |   TestCase |         Mean |          Error |        StdDev |       Median | Scaled | ScaledSD | Allocated |
|-------------------------------------------------- |-------------- |----------- |-------------:|---------------:|--------------:|-------------:|-------:|---------:|----------:|
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicJson** |     **577.2 ns** |      **7.3825 ns** |     **6.5444 ns** |     **578.8 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  BasicJson |     767.5 ns |     10.1751 ns |     9.5178 ns |     772.1 ns |   1.33 |     0.02 |       0 B |
|                                                   |               |            |              |                |               |              |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **HelloWorld** |     **110.0 ns** |      **0.8933 ns** |     **0.7918 ns** |     **110.3 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True | HelloWorld |     311.6 ns |      5.2154 ns |     4.8785 ns |     314.0 ns |   2.83 |     0.05 |       0 B |
|                                                   |               |            |              |                |               |              |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json400B** |     **897.3 ns** |     **17.6848 ns** |    **22.3655 ns** |     **887.9 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |   Json400B |   1,087.4 ns |      9.4323 ns |     7.8764 ns |   1,088.1 ns |   1.21 |     0.03 |       0 B |
|                                                   |               |            |              |                |               |              |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **Json400KB** | **880,193.8 ns** | **10,511.7939 ns** | **9,832.7375 ns** | **881,566.1 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  Json400KB | 868,874.0 ns |  3,904.7297 ns | 3,461.4409 ns | 869,044.8 ns |   0.99 |     0.01 |       0 B |
|                                                   |               |            |              |                |               |              |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json40KB** |  **81,543.4 ns** |  **1,362.8043 ns** | **1,208.0904 ns** |  **81,673.2 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |   Json40KB |  81,089.5 ns |  1,523.3795 ns | 1,496.1623 ns |  80,365.4 ns |   0.99 |     0.02 |       0 B |
|                                                   |               |            |              |                |               |              |        |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |    **Json4KB** |   **6,321.8 ns** |    **108.5949 ns** |    **84.7838 ns** |   **6,358.8 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |    Json4KB |   7,570.4 ns |    156.1703 ns |   403.1254 ns |   7,379.2 ns |   1.20 |     0.07 |       0 B |

|                Method |  TestCase | segmentSize |         Mean |      Error |     StdDev | Allocated |
|---------------------- |---------- |------------ |-------------:|-----------:|-----------:|----------:|
|  **MultiSegmentSequence** | **Json400KB** |        **4000** | **1,192.383 us** | **13.9206 us** | **12.3402 us** |       **0 B** |
| **SingleSegmentSequence** | **Json400KB** |           **?** | **1,121.939 us** | **17.2794 us** | **13.4906 us** |       **0 B** |
|  **MultiSegmentSequence** |  **Json40KB** |        **4000** |   **115.588 us** |  **2.2417 us** |  **2.0969 us** |       **0 B** |
| **SingleSegmentSequence** |  **Json40KB** |           **?** |   **110.189 us** |  **0.3761 us** |  **0.3141 us** |       **0 B** |
|  **MultiSegmentSequence** |   **Json4KB** |        **4000** |     **9.761 us** |  **0.0539 us** |  **0.0477 us** |       **0 B** |
| **SingleSegmentSequence** |   **Json4KB** |           **?** |     **9.310 us** |  **0.2049 us** |  **0.1917 us** |       **0 B** |