@Tornhoof 2018-10-30
Just wondering, what's the upside of this PR, you primarily describe regressions and the changes don't sound they will improve the code quality or similar that much.