@ahsonkhan 2018-10-31
> Just wondering, what's the upside of this PR, you primarily describe regressions and the changes don't sound they will improve the code quality or similar that much.

Cleaner lifetime semantics of the returned ReadOnlySpan/ReadOnlySequence and since there is no longer any need to Dispose, we don't get bugs related to double dispose/memory leak/etc. in the callers code.